Check whether two stan models objects agree on draws in sampling_result.

Check whether two stan models objects agree on draws in sampling_result.

CheckModelEquivalence(sampling_result, model_1, model_2, stan_data_1,
  stan_data_2, check_grads = TRUE, tol = 1e-08, max_chains = 1,
  max_num_samples = 100)

Arguments

sampling_result
Samples drawn from a Stan model, possibly distinct from model_1 and model_2.

model_1
The path to the first stan model.

model_2
The path to the second stan model.

stan_data_1
A stan data file for the first model.

stan_data_2
A stan data file for the second model.

check_grads
Whether to check the gradients of the log probabilities.

tol
The tolerance for the difference in grads and log probability.

Value

A boolean indicating whether model_1 and model_2 agree with each other on the samples specified in sampling_result. Log probability is is compared up to a constant, and the gradients are compared on the parameters in common according to the unconstrained parameter names.