Evaluate a model (represented as a stanfit object) at MCMC draws, possibly from a different model. The model model_stanfit is evaluated at the draws found in samples_stanfit. If there are parameters in model_stanfit that are not contained in samples_stanfit, values from model_par_list are used.

This was designed to evaluate gradients with respect to hyperparameters which are Stan parameters in model_stanfit but not in samples_stanfit. The values of those hyperparameters would be specified in model_par_list and reused for each evaluation of model_stanfit. Since this is the intended use case, the function requires that model_stanfit contains a (non-strict) superset of the parameters in samples_stanfit.

EvaluateModelAtDraws(samples_stanfit, model_stanfit, model_par_list,
  compute_grads = FALSE, max_chains = Inf, max_num_samples = Inf,
  verbose = TRUE)

Arguments

samples_stanfit
(sampling_result) The output of Stan sampling containing the draws at which you want to evaluate the model. (e.g. rstan::sampling())

model_stanfit
(model_fit) A stanfit object from a model, possibly containing a superset of the parameters in samples_stanfit.

model_par_list
A list of parameters for model_stanfit which can be passed to unconstrain_pars for the model_stanfit model. Every parameter in get_inits applied to samples_stanfit must be also be found in model_par_list.

compute_grads
If FALSE, only return the log probability.

max_chains
How many chains to evaluate. The default is to evaluate the number of chains in samples_stanfit.

max_num_samples
How many samples to evaluate. The default is to evaluate the number of non-warmup samples in samples_stanfit.

verbose
If true, display progress messages.

Value

A list with lp_vec containing a vector of log probabilities and, if compute_grads is TRUE, gradients of the log probability, all with resepct to the parameters in model_stanfit at the draws in samples_stanfit. If the sampler contains multiple chains the chains are concatenated in order.